Add query model for category order hints. Move CategoryOrderHints to product package. #737
