+++ sphere-models/src/main/java/io/sphere/sdk/products/ProductDataImpl.java
@@ -2,7 +2,6 @@ package io.sphere.sdk.products;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import io.sphere.sdk.categories.Category;
-import io.sphere.sdk.categories.CategoryOrderHints;
 import io.sphere.sdk.models.Base;
 import io.sphere.sdk.models.LocalizedString;
 import io.sphere.sdk.models.Reference;
