+++ sphere-models/src/main/java/io/sphere/sdk/products/queries/CategoryOrderHintsQueryModel.java
@@ -1,0 +1,7 @@
+package io.sphere.sdk.products.queries;
+
+import io.sphere.sdk.queries.DirectionlessQuerySort;
+
+public interface CategoryOrderHintsQueryModel<T> {
+    DirectionlessQuerySort<T> category(final String categoryId);
+}
