+++ sphere-models/src/main/java/io/sphere/sdk/products/queries/CategoryOrderHintsQueryModelImpl.java
@@ -1,0 +1,16 @@
+package io.sphere.sdk.products.queries;
+
+import io.sphere.sdk.queries.DirectionlessQuerySort;
+import io.sphere.sdk.queries.QueryModel;
+import io.sphere.sdk.queries.QueryModelImpl;
+
+final class CategoryOrderHintsQueryModelImpl<T> extends QueryModelImpl<T> implements CategoryOrderHintsQueryModel<T> {
+    public CategoryOrderHintsQueryModelImpl(final QueryModel<T> parent, final String pathSegment) {
+        super(parent, pathSegment);
+    }
+
+    @Override
+    public DirectionlessQuerySort<T> category(final String categoryId) {
+        return new DirectionlessQuerySort<>(new QueryModelImpl<T>(this, categoryId){});
+    }
+}
